refactor checkpoint
