+++ Clipboard-Manager.xcodeproj/xcuserdata/uriyakir.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -3,20 +3,4 @@
    uuid = "9713723D-F3E2-43EF-897F-F5CFE13DEEAC"
    type = "1"
    version = "2.0">
-   <Breakpoints>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            uuid = "C30A193B-6783-4BC4-8A9E-5E3DCD4E0841"
-            shouldBeEnabled = "Yes"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "Clipboard-Manager/AppDelegate.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "32"
-            endingLineNumber = "32">
-         </BreakpointContent>
-      </BreakpointProxy>
-   </Breakpoints>
 </Bucket>
