+++ Clipboard-Manager/Database/DatabaseHandler.swift
