+++ Clipboard-Manager/VC/MainViewController.swift
